- Updated Babel for decorators (https://babeljs.io/docs/plugins/transform-decorators/) - Updated codeclimate settings for parsed & built files - Updated Yarn and Package.json - Updated folder structures - Updated Auth Components
