+++ .babelrc
@@ -3,6 +3,7 @@
     "next/babel"
   ],
   "plugins": [
+    "transform-decorators-legacy",
     ["styled-components", {
       "ssr": true,
       "displayName": true,
