+++ libraries/reduxStore.js
@@ -2,7 +2,7 @@ import { createStore } from 'redux';
 import getReducer from './reducer';
 import createMiddleware from './middleware';
 import persist from './persist';
-import { AUTH_SIGNIN, AUTH_SIGNOUT } from './auth/authActions';
+import { dispatchers } from '../components/AuthFields/index.data';
 
 let reduxStore = null;
 
@@ -11,17 +11,20 @@ export default (apolloClient, initialState) => {
   if (!process.browser || !reduxStore) {
     const middleware = createMiddleware(apolloClient.middleware());
     store = createStore(getReducer(apolloClient), initialState, middleware);
+
     (async () => {
       const token = await Promise.resolve(persist.willGetAccessToken());
       if (token) {
-        store.dispatch({ type: AUTH_SIGNIN });
+        dispatchers.signIn(token);
       } else if (!token) {
-        store.dispatch({ type: AUTH_SIGNOUT });
+        dispatchers.signOut();
       }
     })();
+
     if (!process.browser) {
       return store;
     }
+
     reduxStore = store;
   }
   return reduxStore;
