+++ package.json
@@ -24,6 +24,7 @@
     "chalk": "2.0.1",
     "color": "2.0.0",
     "compression": "1.7.0",
+    "core-decorators": "0.19.0",
     "cors": "2.8.3",
     "express": "4.15.3",
     "graphql": "0.10.3",
@@ -79,6 +80,7 @@
   },
   "devDependencies": {
     "babel-eslint": "7.2.3",
+    "babel-plugin-transform-decorators-legacy": "1.3.4",
     "cli-clear": "1.0.4",
     "eslint": "3.19.0",
     "eslint-config-airbnb": "15.0.1",
