+++ pages/signin.js
@@ -1,11 +1,9 @@
 import withData from '../libraries/withData';
-import AuthForm from '../containers/AuthForm';
+import SignInForm from '../components/SignInForm';
 import DefaultCon from '../containers/Default';
-
-const SigninContainer = AuthForm.SigninContainer;
 
 export default withData(props =>
   <DefaultCon title="Sign In" {...props}>
-    <SigninContainer selectFields={'signinFields'} />
+    <SignInForm />
   </DefaultCon>
 );
