+++ pages/signup.js
@@ -1,11 +1,9 @@
-import AuthForm from '../containers/AuthForm';
+import SignUpForm from '../components/SignUpForm';
 import withData from '../libraries/withData';
 import DefaultCon from '../containers/Default';
 
-const SignupContainer = AuthForm.SignupContainer;
-
 export default withData(props =>
   <DefaultCon title="Sign Up" {...props}>
-    <SignupContainer selectFields={'signupFields'} />
+    <SignUpForm />
   </DefaultCon>
 );
